fix: 요청에 albumId 없을 시 기본 앨범 사용하도록 구조 수정

+++ src/main/kotlin/kr/co/lokit/api/domain/album/application/AlbumService.kt
@@ -1,5 +1,7 @@
 package kr.co.lokit.api.domain.album.application
 
+import kr.co.lokit.api.common.exception.BusinessException
+import kr.co.lokit.api.common.exception.entityNotFound
 import kr.co.lokit.api.domain.album.domain.Album
 import kr.co.lokit.api.domain.album.infrastructure.AlbumRepository
 import org.springframework.stereotype.Service
@@ -18,10 +20,22 @@ class AlbumService(
         albumRepository.findAllByUserId(userId)
 
     @Transactional
-    fun updateTitle(id: Long, title: String): Album =
-        albumRepository.applyTitle(id, title)
+    fun updateTitle(id: Long, title: String): Album {
+        val album = albumRepository.findById(id)
+            ?: throw entityNotFound<Album>(id)
+        if (album.isDefault) {
+            throw BusinessException.BusinessRuleViolationException("기본 앨범의 이름은 변경할 수 없습니다.")
+        }
+        return albumRepository.applyTitle(id, title)
+    }
 
     @Transactional
-    fun delete(id: Long) =
+    fun delete(id: Long) {
+        val album = albumRepository.findById(id)
+            ?: throw entityNotFound<Album>(id)
+        if (album.isDefault) {
+            throw BusinessException.BusinessRuleViolationException("기본 앨범은 삭제할 수 없습니다.")
+        }
         albumRepository.deleteById(id)
+    }
 }
